Optimize performance for power-of-4 check
Using the cmath functions is far too expensive when simply bitwise operations can be used. Powers of four can be checked by two simple conditions, so I've rewritten the function to check those in a way that it's easy to follow.

This allows us to remove the <cmath> header too, making compilation faster, and I've renamed the function because the word "check" is ambiguous.